db_management_project/requirements.txt
mysql-connector-python==8.0.33
fpdf==1.7.2
python-dotenv==1.0.0